feat: refactor dashboard layout by extracting header and sidebar components
